initial clean up, package names refactor are still missing
