+++ activiti-cloud-services-org/activiti-cloud-services-org-jpa/pom.xml
@@ -80,10 +80,5 @@
       <artifactId>spring-boot-starter-data-jpa</artifactId>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>org.springframework.cloud</groupId>
-      <artifactId>spring-cloud-starter-netflix-ribbon</artifactId>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 </project>
