+++ activiti-cloud-services-org/activiti-cloud-services-org-jpa/src/test/resources/application.properties
@@ -1,0 +1,1 @@
+activiti.cloud.modeling.url=localhost:8088
