+++ activiti-cloud-services-org/activiti-cloud-services-org-rest/src/test/resources/application.properties
@@ -1,1 +1,1 @@
-stubrunner.idsToServiceIds.activiti-cloud-services-process-model-rest=modeling-app
+activiti.cloud.modeling.url=localhost:8088
