+++ activiti-cloud-starter-org/src/test/java/org/activiti/cloud/organization/Application.java
@@ -16,16 +16,12 @@
 
 package org.activiti.cloud.organization;
 
-import org.activiti.cloud.org.config.EnableActivitiOrganization;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.web.client.RestTemplate;
 
 @SpringBootApplication
 @EnableActivitiOrganization
-@ComponentScan("org.activiti")
 public class Application {
 
     public static void main(String[] args) {
