+++ activiti-cloud-starter-org/src/test/resources/application.properties
@@ -13,28 +13,3 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-server.port=0
-spring.application.name=org-service
-keycloak.auth-server-url=http://localhost:8180/auth
-keycloak.realm=springboot
-keycloak.resource=activiti
-keycloak.public-client=true
-
-keycloak.security-constraints[0].authRoles[0]=user
-keycloak.security-constraints[0].securityCollections[0].patterns[0]=/*
-
-keycloak.principal-attribute=preferred-username
-# see https://issues.jboss.org/browse/KEYCLOAK-810 for configuration options
-
-activiti.keycloak.admin-client-app=admin-cli
-activiti.keycloak.client-user=client
-activiti.keycloak.client-password=client
-activiti.keycloak.test-user=testuser
-activiti.keycloak.test-password=password
-
-spring.rabbitmq.host=localhost
-spring.cloud.discovery.enabled=false
-
-
-eureka.client.enabled=false
