--- activiti-cloud-starter-org/src/test/resources/bootstrap.properties
@@ -1,1 +1,0 @@
-spring.cloud.config.enabled=false
